Add doc comments to pyclasses for documentation
These get converted to docstrings in python
and are available through `help()` among others.

However, vscode does not support this. If it was any other editor,
I would not care, but as most students will be using it, I'll have
to find some fix for that at some point to get them to show up.